move <vr-template>'s `utils` and custom `registerElement` call to separate modules
